+++ stroom-index-shared/src/main/java/stroom/index/shared/IndexField.java
@@ -254,20 +254,17 @@ public class IndexField implements HasDisplayValue, Comparable<IndexField>, Seri
             switch (fieldType) {
                 case ID:
                     conditions.add(Condition.EQUALS);
-                    conditions.add(Condition.CONTAINS);
                     conditions.add(Condition.IN);
                     conditions.add(Condition.IN_DICTIONARY);
                     break;
                 case FIELD:
                     conditions.add(Condition.EQUALS);
-                    conditions.add(Condition.CONTAINS);
                     conditions.add(Condition.IN);
                     conditions.add(Condition.IN_DICTIONARY);
                     break;
 
                 case NUMERIC_FIELD:
                     conditions.add(Condition.EQUALS);
-                    conditions.add(Condition.CONTAINS);
                     conditions.add(Condition.GREATER_THAN);
                     conditions.add(Condition.GREATER_THAN_OR_EQUAL_TO);
                     conditions.add(Condition.LESS_THAN);
@@ -279,7 +276,6 @@ public class IndexField implements HasDisplayValue, Comparable<IndexField>, Seri
 
                 case DATE_FIELD:
                     conditions.add(Condition.EQUALS);
-                    conditions.add(Condition.CONTAINS);
                     conditions.add(Condition.GREATER_THAN);
                     conditions.add(Condition.GREATER_THAN_OR_EQUAL_TO);
                     conditions.add(Condition.LESS_THAN);