+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+
+## [v6.0.2] - 2019-08-02
+
 * Issue **#1230** : Fix test.
  
 
@@ -1810,7 +1813,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 * Issue **#202** : Initial release of the new data retention policy functionality.
 
-[Unreleased]: https://github.com/gchq/stroom/compare/v6.0.1...6.0
+[Unreleased]: https://github.com/gchq/stroom/compare/v6.0.2...6.0
+[v6.0.2]: https://github.com/gchq/stroom/compare/v6.0.1...v6.0.2
 [v6.0.1]: https://github.com/gchq/stroom/compare/v6.0-beta.66...v6.0.1
 [v6.0-beta.66]: https://github.com/gchq/stroom/compare/v6.0-beta.65...v6.0-beta.66
 [v6.0-beta.65]: https://github.com/gchq/stroom/compare/v6.0-beta.64...v6.0-beta.65